refactor(MatchStatus): MatchStatus model 폴더로 이동
